tests: add tests for manage_indexer_queues task
