+++ setup.cfg
@@ -31,7 +31,7 @@ install_requires =
     invenio-base>=1.2.11
     invenio-files-rest>=1.3.0
     invenio-i18n>=1.3.2
-    invenio-indexer>=1.2.4
+    invenio-indexer>=1.2.5
     invenio-jsonschemas>=1.1.3
     invenio-pidstore>=1.2.2
     invenio-records-permissions>=0.13.0,<0.14.0
